Rename to `ocm`
This patch renames the tool to `ocm` and adjusts the code for the new
home at `github.com/openshift-online/ocm-cli`.
