+++ .gitignore
@@ -1,1 +1,1 @@
-/uhc
+/ocm
